+++ investing_algorithm_framework/domain/__init__.py
@@ -11,7 +11,7 @@ from .constants import ITEMIZE, ITEMIZED, PER_PAGE, PAGE, ENVIRONMENT, \
     APP_MODE, DATABASE_DIRECTORY_NAME, BACKTESTING_INITIAL_AMOUNT, \
     APPLICATION_DIRECTORY, SNAPSHOT_INTERVAL, AWS_S3_STATE_BUCKET_NAME, \
     LAST_SNAPSHOT_DATETIME, DATA_DIRECTORY, INDEX_DATETIME, \
-    DATETIME_FORMAT_FILE_NAME, DEFAULT_DATETIME_FORMAT
+    DATETIME_FORMAT_FILE_NAME, DEFAULT_DATETIME_FORMAT, TIMEZONE
 from .data_provider import DataProvider
 from .data_structures import PeekableQueue
 from .decimal_parsing import parse_decimal_to_string, parse_string_to_decimal
@@ -48,7 +48,7 @@ from .utils import random_string, append_dict_as_row_to_csv, \
     add_column_headers_to_csv, get_total_amount_of_rows, \
     convert_polars_to_pandas, random_number, is_jupyter_notebook, \
     csv_to_list, StoppableThread, load_csv_into_dict, tqdm, \
-    is_timezone_aware, sync_timezones, get_timezone
+    is_timezone_aware, sync_timezones, get_timezone, format_datetime_utc
 from .backtesting import BacktestRun, BacktestSummaryMetrics, \
     BacktestDateRange, Backtest, BacktestMetrics, combine_backtests, \
     combine_multi_universe_backtest, BacktestEngine, \
@@ -107,6 +107,7 @@ __all__ = [
     "Order",
     "Strategy",
     "DATETIME_FORMAT",
+    "TIMEZONE",
     "StatelessActions",
     "parse_decimal_to_string",
     "parse_string_to_decimal",
@@ -165,6 +166,7 @@ __all__ = [
     "is_timezone_aware",
     "sync_timezones",
     "get_timezone",
+    "format_datetime_utc",
     "Event",
     "SNAPSHOT_INTERVAL",
     "SnapshotInterval",