+++ investing_algorithm_framework/domain/utils/__init__.py
@@ -4,7 +4,8 @@ from .random import random_string, random_number
 from .stoppable_thread import StoppableThread
 from .synchronized import synchronized
 from .polars import convert_polars_to_pandas
-from .dates import is_timezone_aware, sync_timezones, get_timezone
+from .dates import is_timezone_aware, sync_timezones, get_timezone, \
+    format_datetime_utc
 from .jupyter_notebook_detection import is_jupyter_notebook
 from .custom_tqdm import tqdm
 
@@ -22,6 +23,7 @@ __all__ = [
     'is_timezone_aware',
     'sync_timezones',
     'get_timezone',
+    'format_datetime_utc',
     'is_jupyter_notebook',
     'tqdm'
 ]